entirely split maps from lcm
- virtual Polymorphism used to provide API to maps via SensorDataReceiver
- make lcm specific libraries
- makes lcm specific applications
