+++ maps_server/src/apps/CMakeLists.txt
@@ -2,7 +2,7 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g -std=c++0x -Wno-deprecated-declaratio
 
 add_executable(data-request-server DataRequestServer.cpp RobotState.cpp)
 target_link_libraries(data-request-server pthread)
-pods_use_pkg_config_packages(data-request-server maps)
+pods_use_pkg_config_packages(data-request-server maps-lcm)
 pods_install_executables(data-request-server)
 
 add_executable(maps-test TestMain.cpp)
@@ -10,16 +10,16 @@ pods_use_pkg_config_packages(maps-test maps bot2-lcmgl-client)
 pods_install_executables(maps-test)
 
 add_executable(maps-test-collector TestCollector.cpp)
-pods_use_pkg_config_packages(maps-test-collector maps bot2-lcmgl-client maps_lcm)
+pods_use_pkg_config_packages(maps-test-collector bot2-lcmgl-client maps-lcm)
 pods_install_executables(maps-test-collector)
 
 add_executable(maps-snapshot-server SnapshotServer.cpp)
-pods_use_pkg_config_packages(maps-snapshot-server maps maps_lcm)
+pods_use_pkg_config_packages(maps-snapshot-server maps-lcm)
 pods_install_executables(maps-snapshot-server)
 
 add_executable(maps-registrar PointCloudRegistrar.cpp)
 target_link_libraries(maps-registrar pthread)
-pods_use_pkg_config_packages(maps-registrar maps bot2-lcmgl-client)
+pods_use_pkg_config_packages(maps-registrar bot2-lcmgl-client maps-lcm)
 pods_install_executables(maps-registrar)
 
 #add_executable(maps-affordance-updater AffordanceUpdater.cpp)
