+++ maps_server/src/maps_lcm_utils/BotWrapper.hpp
