+++ maps_server/src/maps_lcm_utils/CMakeLists.txt
@@ -1,0 +1,32 @@
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++0x -Wall -Wno-deprecated-declarations")
+
+
+# Create a shared library libmaps-lcm-utils.so
+add_library(maps-lcm-utils SHARED
+    Clock.cpp
+    LcmWrapper.cpp
+    BotWrapper.cpp
+)
+
+# make the header public
+# install it to include/maps_utils
+pods_install_headers(
+  Clock.hpp
+  LcmWrapper.hpp
+  BotWrapper.hpp
+  DESTINATION maps_lcm_utils)
+
+# make the library public
+pods_install_libraries(maps-lcm-utils)
+
+# uncomment these lines to link against another library via pkg-config
+set(REQUIRED_PACKAGES bot2-core lcm eigen3 bot2-frames maps-utils)
+pods_use_pkg_config_packages(maps-lcm-utils ${REQUIRED_PACKAGES})
+
+# create a pkg-config file for the library, to make it easier for other
+# software to use.
+pods_install_pkg_config_file(maps-lcm-utils
+    CFLAGS
+    LIBS -lmaps-lcm-utils
+    REQUIRES ${REQUIRED_PACKAGES}
+    VERSION 0.0.1)
