+++ maps_server/src/maps_utils/CMakeLists.txt
@@ -3,16 +3,12 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++0x -Wall -Wno-deprecated-declara
 
 # Create a shared library libmaps-utils.so
 add_library(maps-utils SHARED
-    Clock.cpp
-    LcmWrapper.cpp
     LidarUtils.cpp
 )
 
 # make the header public
 # install it to include/maps_utils
 pods_install_headers(
-  Clock.hpp
-  LcmWrapper.hpp
   LidarUtils.hpp
   PointerUtils.hpp
   ThreadSafeQueue.hpp
@@ -32,33 +28,5 @@ pods_use_pkg_config_packages(maps-utils ${REQUIRED_PACKAGES})
 pods_install_pkg_config_file(maps-utils
     CFLAGS
     LIBS -lmaps-utils
-    REQUIRES ${REQUIRED_PACKAGES}
-    VERSION 0.0.1)
-
-
-
-# Create a shared library libmaps-bot-wrapper.so
-add_library(maps-bot-wrapper SHARED
-    BotWrapper.cpp
-)
-
-# make the header public
-# install it to include/maps_utils
-pods_install_headers(
-  BotWrapper.hpp
-  DESTINATION maps_bot_wrapper)
-
-# make the library public
-pods_install_libraries(maps-bot-wrapper)
-
-# uncomment these lines to link against another library via pkg-config
-set(REQUIRED_PACKAGES bot2-core lcm eigen3 bot2-frames maps-utils)
-pods_use_pkg_config_packages(maps-bot-wrapper ${REQUIRED_PACKAGES})
-
-# create a pkg-config file for the library, to make it easier for other
-# software to use.
-pods_install_pkg_config_file(maps-bot-wrapper
-    CFLAGS
-    LIBS -lmaps-bot-wrapper
     REQUIRES ${REQUIRED_PACKAGES}
     VERSION 0.0.1)
