+++ maps_server/src/server/StereoHandler.cpp
@@ -3,7 +3,7 @@
 #include <maps/DepthImage.hpp>
 #include <maps/DepthImageView.hpp>
 #include <maps/Utils.hpp>
-#include <maps_bot_wrapper/BotWrapper.hpp>
+#include <maps_lcm_utils/BotWrapper.hpp>
 
 #include <lcm/lcm-cpp.hpp>
 #include <lcmtypes/bot_core/images_t.hpp>
